// utils/roles-enum.ts
export enum ROLES {
  DOCTOR = "Doctor",
  PATIENT = "Patient",
  NURSE = "Nurse",
  PHARMACY = "Pharmacy",
  RECEPTION = "Reception",
  HOSPITAL = "Hospital",
}

export enum STORAGE_KEYS {
  ROLE = "role",
  DOCTOR_ID = "doctorId",
  HOSPITAL_ID = "hospitalId",
  TOKEN = "token",
  IS_LEAD = "isLead",
}